Include License file
You mention in the README that this is an open-source database. This includes a license that makes your software truly open source. This repository is probably not being updated, but for the many developers that are using taffydb as a dependency in an enterprise environment, this makes taffydb appear as a license risk.